requirements.txt
accelerate==1.9.0
adagio==0.2.6
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.3
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
autocommand==2.2.2
av==15.0.0
babel==2.17.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
bps_torch @ git+https://github.com/otaheri/bps_torch@00fcac3dffd294211d4ea0363eaf6c2ca4e068dc
braceexpand==0.1.7
brotli==1.2.0
certifi==2025.7.9
cffi==1.17.1
cfgv==3.5.0
chamfer_distance @ git+https://github.com/otaheri/chamfer_distance@f86f6f7cadd3aca642704573d1626c67ca2e2846
charset-normalizer==3.4.2
chumpy @ git+https://github.com/mattloper/chumpy@9b045ff5d6588a24a0bab52c83f032e2ba433e17
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.2
coreforecast==0.0.16
cycler==0.12.1
Cython==3.1.2
darts==0.36.0
dash==3.1.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.18
diffusers==0.36.0
distlib==0.4.0
distro==1.9.0
e3nn==0.5.9
einops==0.8.1
entrypoints==0.4
evo==1.34.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.126.0
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
ffmpy==1.0.0
filelock==3.13.1
flash_attn==2.4.3.post1
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.61.1
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.7.0
fs==2.4.16
fsspec==2024.6.1
fugue==0.9.4
future==1.0.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
GPUtil==1.4.0
gradio==6.2.0
gradio_client==2.0.2
groovy==0.1.2
h11==0.16.0
h5py==3.14.0
hdf5plugin==5.1.0
hf-xet==1.2.0
holidays==0.78
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-core==1.3.2
hydra_zen==0.16.0
identify==2.6.15
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
inflect==7.3.1
iopath==0.1.10
ipycanvas==0.13.3
ipyevents==2.0.2
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jaxtyping==0.3.4
jedi==0.19.2
Jinja2==3.1.4
jiter==0.12.0
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
lightning-utilities==0.15.2
llvmlite==0.44.0
loguru==0.7.3
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.46.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nfoursid==1.0.2
ninja==1.11.1.4
nodeenv==1.10.0
notebook_shim==0.2.4
numba==0.61.2
numexpr==2.14.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
triton==3.2.0
objaverse==0.1.7
omegaconf==2.3.0
OneEuroFilter==0.2.1
onnx==1.18.0
onnxruntime-gpu==1.22.0
open3d==0.19.0
openai==2.14.0
opencv-contrib-python==4.12.0.88
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
opt-einsum-fx==0.1.4
opt_einsum==3.4.0
optional-django==0.1.0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
pillow_heif==1.1.0
platformdirs==4.3.8
plotly==6.2.0
plyfile==1.1.3
portalocker==3.2.0
pre_commit==4.5.1
proglog==0.1.12
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pybind11==3.0.0
pycolmap==3.13.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyglet==2.1.6
pygltflib==1.16.4
Pygments==2.19.2
pyod==2.0.5
PyOpenGL==3.1.0
pyparsing==3.2.3
pyquaternion==0.9.9
pyrender==0.1.45
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-multipart==0.0.21
pytorch-lightning==2.5.2
PyTurboJPEG @ git+https://github.com/lilohuang/PyTurboJPEG.git@ef7660a0e738f49a33d0a144d1e573c7f4eabfcc
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2025.11.3
requests==2.32.4
retrying==1.4.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.2.0
rosbags==0.11.0
rpds-py==0.26.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safehttpx==0.1.7
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==2.34.1
shap==0.48.0
shapely==2.1.1
shellingham==1.5.4
simsimd==6.5.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
smplx==0.1.28
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.50.0
statsforecast==2.0.2
statsmodels==0.14.5
stringzilla==3.12.5
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.4
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.17
tinycss2==1.4.0
tokenizers==0.22.1
tomli==2.2.1
tomlkit==0.13.3
torchgeometry==0.1.2
torchmetrics==1.8.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformations==2025.1.1
transformers==4.57.6
triad==1.0.0
trimesh==4.7.0
typeguard==4.3.0
typer==0.20.1
typer-slim==0.20.1
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.6.3
usd-core==25.5.1
utilsforecast==0.2.12
uvicorn==0.38.0
videoio==0.3.0
virtualenv==20.35.4
wadler_lindig==0.1.7
wandb==0.21.0
warp-lang==1.8.0
wcwidth==0.2.13
webcolors==24.11.1
webdataset==1.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xarray==2025.6.1
xformers==0.0.29.post3
xgboost==3.0.3
yacs==0.1.8
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
smplfitter
git+https://github.com/imaug/imaug.git # compatible with np>2
git+https://github.com/facebookresearch/pytorch3d.git
git+https://github.com/NVIDIAGameWorks/kaolin.git
git+https://github.com/NVlabs/nvdiffrast.git
